chore: 패키지 구조 세분화 및 패키지 이름 오타 수정
- sevice -> service
