+++ src/main/java/janggi/Application.java
@@ -5,8 +5,8 @@ import janggi.repository.util.ConnectionProvider;
 import janggi.repository.util.DBConnectionProvider;
 import janggi.repository.util.SchemaInitializer;
 import janggi.repository.util.TransactionManager;
-import janggi.sevice.DefaultJanggiService;
-import janggi.sevice.JanggiService;
+import janggi.service.DefaultJanggiService;
+import janggi.service.JanggiService;
 import janggi.view.ApplicationView;
 import janggi.view.input.ConsoleReader;
 import janggi.view.output.ConsoleWriter;
