+++ src/main/java/janggi/domain/strategy/arrangement/MaSangSangMa.java
@@ -1,15 +1,15 @@
-package janggi.domain.strategy;
+package janggi.domain.strategy.arrangement;
 
 import janggi.domain.Side;
 import janggi.domain.piece.Ma;
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.Sang;
 
-public class MaSangMaSang implements ArrangementStrategy {
+public class MaSangSangMa implements ArrangementStrategy {
 
     private final Side side;
 
-    public MaSangMaSang(Side side) {
+    public MaSangSangMa(Side side) {
         this.side = side;
     }
 
@@ -19,7 +19,7 @@ public class MaSangMaSang implements ArrangementStrategy {
         int row = calculateRow(boardMaxLength, side);
         board[row][1] = new Ma(side);
         board[row][2] = new Sang(side);
-        board[row][6] = new Ma(side);
-        board[row][7] = new Sang(side);
+        board[row][6] = new Sang(side);
+        board[row][7] = new Ma(side);
     }
 }
