+++ src/main/java/janggi/service/DefaultJanggiService.java
@@ -1,21 +1,21 @@
-package janggi.sevice;
+package janggi.service;
 
-import janggi.domain.Location;
+import janggi.domain.board.Location;
 import janggi.domain.Side;
 import janggi.domain.board.Board;
 import janggi.domain.piece.Piece;
 import janggi.domain.state.GameContext;
-import janggi.domain.strategy.ArrangementStrategy;
+import janggi.domain.strategy.arrangement.ArrangementStrategy;
 import janggi.domain.strategy.BoardAssembler;
-import janggi.domain.strategy.DBArrangementStrategy;
-import janggi.domain.strategy.IntersectionInitializer;
+import janggi.domain.strategy.arrangement.DBArrangementStrategy;
+import janggi.domain.strategy.intersection.IntersectionInitializer;
 import janggi.repository.dao.GameDao;
 import janggi.repository.dao.PieceDao;
 import janggi.repository.entity.GameEntity;
 import janggi.repository.entity.PieceEntity;
 import janggi.repository.util.TransactionManager;
-import janggi.sevice.dto.GameInformation;
-import janggi.sevice.mapper.PieceEntityMapper;
+import janggi.service.dto.GameInformation;
+import janggi.service.mapper.PieceEntityMapper;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
