+++ src/test/java/janggi/domain/rule/route/RouteTest.java
@@ -1,6 +1,6 @@
 package janggi.domain.rule.route;
 
-import janggi.domain.Location;
+import janggi.domain.board.Location;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
