+++ src/test/java/janggi/repository/dao/PieceDaoTest.java
@@ -1,7 +1,7 @@
 package janggi.repository.dao;
 
 
-import janggi.domain.Location;
+import janggi.domain.board.Location;
 import janggi.repository.entity.GameEntity;
 import janggi.repository.entity.PieceEntity;
 import janggi.repository.util.TransactionManager;
