+++ src/test/java/janggi/service/DefaultJanggiServiceTest.java
@@ -1,21 +1,21 @@
-package janggi.sevice;
+package janggi.service;
 
-import janggi.domain.Location;
+import janggi.domain.board.Location;
 import janggi.domain.Side;
 import janggi.domain.piece.Ma;
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.PieceType;
 import janggi.domain.state.GameContext;
-import janggi.domain.strategy.ArrangementStrategy;
-import janggi.domain.strategy.IntersectionInitializer;
-import janggi.domain.strategy.MaSangMaSang;
-import janggi.domain.strategy.PalaceIntersectionInitializer;
+import janggi.domain.strategy.arrangement.ArrangementStrategy;
+import janggi.domain.strategy.intersection.IntersectionInitializer;
+import janggi.domain.strategy.arrangement.MaSangMaSang;
+import janggi.domain.strategy.intersection.PalaceIntersectionInitializer;
 import janggi.repository.dao.GameDao;
 import janggi.repository.dao.PieceDao;
 import janggi.repository.entity.GameEntity;
 import janggi.repository.entity.PieceEntity;
 import janggi.repository.util.TransactionManager;
-import janggi.sevice.dto.GameInformation;
+import janggi.service.dto.GameInformation;
 import janggi.support.TestArrangementStrategy;
 import janggi.support.TestDBConnectionProvider;
 import java.util.List;
